TW-898: move private class to another file
