+++ lib/pages/chat/events/message/message.dart
@@ -624,321 +624,4 @@ class Message extends StatelessWidget {
       ),
     );
   }
-
-  // Check if the sender of the current event is the same as the compared event.
-  bool isSameSender(Event? comparedEvent, Event currentEvent) {
-    // If the compared event is null, it is assumed that the message is the newest.
-    if (comparedEvent == null) {
-      return true;
-    }
-
-    final isPreviousOrNextEventMessage = {
-      EventTypes.Message,
-      EventTypes.Sticker,
-      EventTypes.Encrypted,
-      EventTypes.Redaction,
-    }.contains(comparedEvent.type);
-
-    // Ignoring events that are not messages, stickers, encrypted or redaction.
-    if (!isPreviousOrNextEventMessage) {
-      return true;
-    }
-
-    return currentEvent.senderId != comparedEvent.senderId;
-  }
-
-  TextPainter _getSizeDisplayName(
-    BuildContext context,
-    double maxWidth,
-  ) {
-    return TextPainter(
-      text: TextSpan(
-        text: event.senderFromMemoryOrFallback
-            .calcDisplayname()
-            .shortenDisplayName(
-              maxCharacters: maxCharactersDisplayNameBubble,
-            ),
-        style: Theme.of(context).textTheme.labelMedium?.copyWith(
-              color: Theme.of(
-                context,
-              ).colorScheme.primary,
-            ),
-      ),
-      maxLines: 2,
-      textDirection: TextDirection.ltr,
-    )..layout(minWidth: 0, maxWidth: maxWidth);
-  }
-
-  TextPainter _getSizeMessageText(
-    BuildContext context,
-    double maxWidth,
-  ) {
-    return TextPainter(
-      text: TextSpan(
-        text: event.calcLocalizedBodyFallback(
-          MatrixLocals(L10n.of(context)!),
-          hideReply: true,
-        ),
-        style: Theme.of(context).textTheme.labelMedium?.copyWith(
-              color: Theme.of(
-                context,
-              ).colorScheme.primary,
-            ),
-      ),
-      textDirection: TextDirection.ltr,
-    )..layout(minWidth: 0, maxWidth: maxWidth);
-  }
-
-  double? _getSizeMessageBubbleWidth(
-    BuildContext context, {
-    required double maxWidth,
-    bool ownMessage = false,
-    bool hideDisplayName = false,
-  }) {
-    final isNotSupportCalcSize = {
-      MessageTypes.File,
-      MessageTypes.Image,
-      MessageTypes.Video,
-    }.contains(event.messageType);
-
-    if (ownMessage || hideDisplayName || isNotSupportCalcSize) {
-      return null;
-    }
-
-    final sizeWidthDisplayName = _getSizeDisplayName(
-      context,
-      maxWidth,
-    ).width;
-
-    final sizeWidthMessageText = _getSizeMessageText(
-      context,
-      maxWidth,
-    ).width;
-
-    return max<double>(sizeWidthDisplayName, sizeWidthMessageText);
-  }
-}
-
-class _DisplayNameWidget extends StatelessWidget {
-  const _DisplayNameWidget({
-    required this.event,
-    required this.maxCharactersDisplayNameBubble,
-  });
-
-  final Event event;
-  final int maxCharactersDisplayNameBubble;
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<User?>(
-      future: event.fetchSenderUser(),
-      builder: (context, snapshot) {
-        final displayName = snapshot.data?.calcDisplayname() ??
-            event.senderFromMemoryOrFallback.calcDisplayname();
-        return Padding(
-          padding: EdgeInsets.only(
-            left: event.messageType == MessageTypes.Image ? 0 : 8.0,
-            bottom: 4.0,
-          ),
-          child: Text(
-            displayName.shortenDisplayName(
-              maxCharacters: maxCharactersDisplayNameBubble,
-            ),
-            style: Theme.of(context).textTheme.labelMedium?.copyWith(
-                  color: Theme.of(
-                    context,
-                  ).colorScheme.primary,
-                ),
-            maxLines: 2,
-            overflow: TextOverflow.clip,
-          ),
-        );
-      },
-    );
-  }
-}
-
-class _ReplyContent extends StatelessWidget {
-  const _ReplyContent({
-    required this.event,
-    required this.timeline,
-    required this.scrollToEventId,
-    required this.ownMessage,
-    required this.controller,
-  });
-
-  final Event event;
-  final Timeline timeline;
-  final void Function(String p1)? scrollToEventId;
-  final bool ownMessage;
-  final ChatController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<Event?>(
-      future: event.getReplyEvent(
-        timeline,
-      ),
-      builder: (
-        BuildContext context,
-        snapshot,
-      ) {
-        final replyEvent = snapshot.data ??
-            Event(
-              eventId: event.relationshipEventId!,
-              content: {
-                'msgtype': 'm.text',
-                'body': '...',
-              },
-              senderId: event.senderId,
-              type: 'm.room.message',
-              room: event.room,
-              status: EventStatus.sent,
-              originServerTs: DateTime.now(),
-            );
-        return InkWell(
-          onTap: () {
-            if (scrollToEventId != null) {
-              scrollToEventId!(
-                replyEvent.eventId,
-              );
-            }
-          },
-          child: AbsorbPointer(
-            child: Container(
-              margin: EdgeInsets.symmetric(
-                vertical: 4.0 * AppConfig.bubbleSizeFactor,
-              ),
-              child: ReplyContent(
-                replyEvent,
-                ownMessage: ownMessage,
-                timeline: timeline,
-                chatController: controller,
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
-
-class ReplyIconWidget extends StatelessWidget {
-  final bool isOwnMessage;
-
-  const ReplyIconWidget({
-    Key? key,
-    required this.isOwnMessage,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Flexible(
-      flex: MessageStyle.replyIconFlexMobile,
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        mainAxisAlignment:
-            isOwnMessage ? MainAxisAlignment.end : MainAxisAlignment.start,
-        children: [
-          if (isOwnMessage)
-            const SizedBox(
-              width: 8.0,
-            ),
-          Transform(
-            alignment: Alignment.center,
-            transform: Matrix4.rotationY(math.pi),
-            child: Container(
-              decoration: BoxDecoration(
-                shape: BoxShape.circle,
-                color: MessageStyle.forwardColorBackground(context),
-              ),
-              width: MessageStyle.forwardContainerSize,
-              height: MessageStyle.forwardContainerSize,
-              child: const Icon(Icons.reply),
-            ),
-          ),
-          if (!isOwnMessage)
-            const SizedBox(
-              width: 12.0,
-            ),
-        ],
-      ),
-    );
-  }
-}
-
-class _MultiPlatformsMessageContainer extends StatelessWidget {
-  final void Function()? onTap;
-
-  final Function(bool) onHover;
-
-  final Widget child;
-
-  const _MultiPlatformsMessageContainer({
-    required this.onTap,
-    required this.onHover,
-    required this.child,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    if (PlatformInfos.isWeb) {
-      return MouseRegion(
-        child: child,
-        onHover: (event) {
-          onHover(true);
-        },
-      );
-    }
-    return InkWell(
-      hoverColor: Colors.transparent,
-      onTap: onTap,
-      onHover: (hover) {
-        onHover(hover);
-      },
-      child: child,
-    );
-  }
-}
-
-class _MultiPlatformSelectionMode extends StatelessWidget {
-  final bool longPressSelect;
-
-  final void Function(Event event)? onSelect;
-
-  final Event event;
-
-  final Widget child;
-
-  final bool useInkWell;
-
-  const _MultiPlatformSelectionMode({
-    required this.longPressSelect,
-    this.onSelect,
-    this.useInkWell = true,
-    required this.event,
-    required this.child,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    if (!useInkWell) {
-      return child;
-    }
-    return Material(
-      color: Colors.transparent,
-      borderRadius: MessageStyle.bubbleBorderRadius,
-      borderOnForeground: false,
-      child: InkWell(
-        onTap: () => onSelect?.call(event),
-        onLongPress: !longPressSelect ? null : () => onSelect?.call(event),
-        borderRadius: MessageStyle.bubbleBorderRadius,
-        hoverColor: Colors.transparent,
-        focusColor: Colors.transparent,
-        splashColor: Colors.transparent,
-        highlightColor: Colors.transparent,
-        child: child,
-      ),
-    );
-  }
 }
